refactor the mock data for assets to keep things DRY
